Memperbarui kode

+++ resources/js/Components/TableMasuk.jsx
@@ -8,12 +8,12 @@ import {
   TableRow,
 } from '@/Components/ui/table';
 import { Button } from './ui/button';
-import { PencilIcon, Trash } from 'lucide-react';
+import { ArrowLeft, PencilIcon, Trash } from 'lucide-react';
 import { Link, router } from '@inertiajs/react';
 import Container from './Container';
 import { formatDate } from 'date-fns';
-export default function TableMasuk({ stocks, params }) {
-  console.table('test nya ', stocks);
+import { AlertDialog, AlertDialogCancel, AlertDialogContent, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger } from './ui/alert-dialog';
+export default function TableMasuk({ stockins, params, setParams }) {
 
   const deleteStockIn = (id) => {
     location.href = `/stock-in?start_date=${params.start_date}&end_date=${params.end_date}`;
@@ -21,8 +21,8 @@ export default function TableMasuk({ stocks, params }) {
   };
   return (
     <>
-      
-      <Table className="border">
+
+      <Table className="border text-xs">
         <TableHeader className="bg-clr-secondary">
           <TableRow>
             <TableHead className="w-[150px]">Tanggal</TableHead>
@@ -32,7 +32,7 @@ export default function TableMasuk({ stocks, params }) {
           </TableRow>
         </TableHeader>
         <TableBody>
-          {stocks.map((stock) => (
+          {stockins.data.map((stock) => (
             <TableRow key={stock.id}>
               <TableCell className="font-medium">
                 {formatDate(new Date(stock.created_at), 'MM/dd/yyyy')}
@@ -49,33 +49,59 @@ export default function TableMasuk({ stocks, params }) {
                   </Button>
                 </Link>
 
-                <Button
-                  onClick={() => deleteStockIn(stock.id)}
-                  variant="ghost"
-                  className="flex items-center gap-1 text-red-500">
-                  <span>
-                    <Trash width={16} height={16} />
-                  </span>
-                  <span>Hapus</span>
-                </Button>
+
+                <AlertDialog>
+                  <AlertDialogTrigger asChild>
+                    <Button
+                      variant="ghost"
+                      className="flex items-center gap-2 text-red-500">
+                      <span>
+                        <Trash width={16} height={16} />
+                      </span>
+                      <span>Hapus</span>
+                    </Button>
+                  </AlertDialogTrigger>
+                  <AlertDialogContent>
+                    <AlertDialogHeader>
+                      <AlertDialogTitle>
+                        Apakah anda benar-benar ingin Menghapus stock
+                        {stock.name}?
+                      </AlertDialogTitle>
+                    </AlertDialogHeader>
+                    <AlertDialogFooter>
+                      <AlertDialogCancel>
+                        <Button variant="ghost" className="w-full">
+                          <div className="flex items-center gap-2 ">
+                            <span>
+                              <ArrowLeft width={17} height={17} />
+                            </span>
+                            <span>Kembali</span>
+                          </div>
+                        </Button>
+                      </AlertDialogCancel>
+                      <AlertDialogFooter>
+                        <AlertDialogCancel>
+                          <Button
+                            variant="ghost"
+                            onClick={() => deleteStockIn(stock.id)}
+                            className="flex items-center gap-2 text-red-500">
+                            <span>
+                              <Trash width={16} height={16} />
+                            </span>
+                            <span>Hapus saja</span>
+                          </Button>
+                        </AlertDialogCancel>
+                      </AlertDialogFooter>
+                    </AlertDialogFooter>
+                  </AlertDialogContent>
+                </AlertDialog>
               </TableCell>
             </TableRow>
           ))}
         </TableBody>
       </Table>
-      {/* <Container className=" relative bottom-0 flex flex-wrap items-center gap-2 w-full">
-        {stocks.meta.links.map((item) => (
-          // <Link
-          //     disabled={item.url === null ? true : false}
-          //     as="button"
-          //     className={`${
-          //         item.url == null
-          //             ? "text-gray-300 cursor-not-allowed"
-          //             : "text-slate-800"
-          //     } w-auto p-2 h-9 flex justify-center items-center border hover:bg-clr-primary transition-all rounded bg-white`}
-          //     href={item.url || ""}
-          //     dangerouslySetInnerHTML={{ __html: item.label }}
-          // />
+      <Container className=" relative bottom-0 flex flex-wrap items-center gap-2 w-full">
+        {stockins.meta.links.map((item) => (
           <Button
             variant="ghost"
             className={`${
@@ -92,7 +118,7 @@ export default function TableMasuk({ stocks, params }) {
             <span dangerouslySetInnerHTML={{ __html: item.label }} />
           </Button>
         ))}
-      </Container> */}
+      </Container>
     </>
   );
 }
